requirements_sol.txt
azure-storage-blob==12.24.0
azure-identity==1.19.0
azure-keyvault==4.2.0
ipykernel==6.29.5
langchain==0.3.8
langchain-community==0.3.8
langchain-postgres==0.0.12
langchain_openai==0.2.9
openpyxl==3.1.5
openai==1.55.0
pandas==2.2.3
postgres==4.0
python-dotenv==1.0.1
tiktoken==0.8.0
httpx==0.27.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
psycopg2-pool==1.2
tqdm==4.67.1